Add LICENSE and README files
This commit introduces a LICENSE file that dictates the terms under which the code can be used. Additionally, a README file has been added to provide essential information about the project, including a project description, setup instructions, usage instructions, a brief on how the application works, and acknowledgment of used resources.